support: add environment controllable feature flags
RFC via PR! This is a mechanism that we have explicitly built into some
of the work for the next release, so I decided to come with a fully
concrete proposal.

Here we provide the macro "KATANA_EXPERIMENTAL_FEATURE(FeatureName)"
where FeatureName is some feature-specific string. Other code can then
then check if the flag using the macro
KATANA_EXPERIMENTAL_ENABLED(FeatureName).

Those wishing to use the experimental feature simply put "FeatureName"
into a comma delimited list provided via the environment variable
"KATANA_ENABLE_EXPERIMENTAL". If "FeatureName" is part of that string
the above function will return true.

The mechanism is very simple on purpose; rather than allow flags that
are "True" by default I thought it was a good idea to encourage
developers to eliminate feature flags once the features they guard are
stable or turn feature flags into actual user facing configuration
options if the ability to toggle a flag is actually useful.

The main thing I like about this design:
 * flags can be defined close to to features they guard
 * related to the above: adding/changing a flag does not require a
   cross-repo commit
 * flags can be as private or as public as they need to be

This comes with the drawback that technically flags can conflict,
meaning a developer could accidentally enable some far flung feature
when they only meant to test their own. I don't think that's a huge
concern as the macro is easy to grep for.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>