libgraph: Adds fixed-binary POD array property and a view over the array
Adds a new property type to Properties.h: FixedSizeBinaryPODArray.
This property is an array on every single node of some fixed size and is
represented with a FixedSizeBinary arrow array. Unlike the existing
ArrayProperty, it does not use arrow Lists (which add overhead to the
existing ArrayProperty since offsets are maintained even though the
ArrayProperty currently only uses fixed sized lists) and supports
arbitrary POD types. This commit also adds a view over
the arrow FixedSizeBinaryArray type assuming it contains some fixed
number of POD types.

The test this commit adds makes sure that this new view is sane.