support: add a new id type for indexing (supports prefix sum)
This is a coercion resistant wrapper for an index. They support index-y
things like being added together an subtracted, but their type keeps
you from using them in the wrong place.

The protections against misuse aren't as strong as the other Opaque
types. IndexIDs are mainly useful for cases when the keys are literally
indexes to something dense like an array, in these cases it's convenient
to do prefix sums over indexes (for instance) when assigning things in
blocks.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>