Prefer strict channel priority
A recent release of the readline library 8.1.2 in the conda channel
pkgs/main highlighted an issue with how conda selects libraries. Under
the default ("flexible") policy, newer versions are preferred regardless
of which channel they come from.

This can cause issues because the pkgs/main channel (the typically
slower moving channel) uses a different compiler toolchain from
the conda-forge channel (the faster moving channel that we want to take
our libraries from).

We were saved from this behavior because most conda-forge packages are
newer than pkgs/main counterparts, but this is not guaranteed in any
way.

Instead, use the "strict" policy, which searches channels in priority
order and does not move to a lower priority channel unless no satisfying
version exists in a higher priority channel. This prefers conda-forge
packages over pkgs/main packages (if a satisfying version exists in
conda-forge).

The strict policy will be the default in Conda 5.0.

Unfortunately, strict priority cannot be applied throughout as some
callers require the flexible policy to be satisfiable. Call out those
exceptions explicitly and defer fixing callers to future work.