support: Encode URIs
Non-alphanumeric characters are usually percent-escaped in URIs. Escape
these characters when converting a URI into a string.

Uri::path() continues to return raw bytes under the assumption that
callers expect it to behave like a standard filesystem path (with no
special escaping).

Along the way remove Uri::Encode, which base64-encodes a URI. This is a
non-standard thing to do to URIs specifically and is not currently used
by any code.