RDG: add NUMAArrray type ids interface
A common pattern is to load the file associated with a type id
array into a FileView buffer and then copy it into a NUMAArray.
Move this pattern into the RDG
itself so it isn't duplicated so many times. Centralizing this in
RDGCore also creates a single location where storage format
complications are handled.