Promote graph projection to a subclass of PropertyGraph (#820)
This PR removes graph projections from the topology cache of a PropertyGraph and makes it a subclass of PropertyGraph instead. This way one can compose projections with other graph views (transposed, undirected, etc).

Note that projections share the RDG instance with their parent property graph. At this point, it is NOT safe to call RDG related methods on the projection, because the projected topologies will conflict with unprojected ones. This will be addressed in the subsequent PR.