Memory API refinement and make sure property caching happens (#890)
* More access to cache statistics

* Memory API restructure

The MemorySupervisor now creates the managers and indexes them by name.
The property cache is no longer exposed and is only accessed indirectly
via the property manager (which itself is accessed via the memory
supervisor).

AddProperties got a Boolean flag because we use it to load properties
and other graph metadata (like the local_to_global array).  I need the
flag to do accurate bookkeeping for property loading.

This makes the ProgressTracer mandatory for more tests. I like the
Progress tracer.