tsuba: Creating RDG interface for python out-of-core import (#859)
Previously, out-of-core import produced v1 RDGs which had been difficult to maintain due to its lack of efficient type lookups. This PR makes the appropriate changes needed for out-of-core import to produce RDGs at the latest version. The major changes needed were the inclusion of the EntityTypeManager as well as a wrapper interface to write out RDG part headers. It adds safe guards to prevent the out-of-core code from going out of sync.