tsuba: move EntityTypeManager out of support
This is a prerequisite for moving away from slow bitset but I haven't
done that yet because the code has drifted such that DynamicBitset is
not a drop-in replacement any longer. I gave up after my effort met some
resistance in serialization.

The real change that I wanted here was to expose these functions that
interface between EntityTypeManager and storage.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>