experimental: switch to global variables and a string registry
At first, we avoided this method because I wanted to prioritize failing
to compile when the developer mistyped a feature flag. That caused some
problems with multiple declarations that we decided were more important.

This version will cause a debug build to fail when a developer uses a
flag that was not registered.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>