Fix copying a projected PG detached from original properties [KAT-4107] (#931)
Clustering operations require copying a property graph to the new purely in-memory instance that only contains the topology. To do it correctly, we have to discard any property index indirection when copying the topology.

I also added debug-mode bounds checking to TypedPropertyGraph and TypedPropertyGraphView methods, which would have caught this issue.