libgalois: fix DynamicBitset count/not bug
DynamicBitset::bitwise_not() applies bitwise not to every element in
DynamicBitset's internal vector. DynamicBitset::count() calls popcount
on every element in DynamicBitset's internal vector and takes a sum.

The last element in the vector has some unused bits in it if the size of
the bitset is not a multiple of 64. When the vector is constructed,
these bits are set to 0. So calling popcount on that element gives only
the set and used bits. But once bitwise_not has been called, those
unused bits are set to 1. And calling popcount on the last element gives
the number of set and used bits plus the number of unused bits.

One solution is to change count to not count the unused bits. One
solution is to change bitwise_not to not invert the unused bits. Change
the behavior of not to maintain the invariant that the unused bits are
always 0 because that gives nice semantics when resizing the bitset.