DynamicBitset: optimizes set-bit getting function, adds test
Getting a list of set bits in a dynamic bitset can be useful; the
implementation of it before this commit did a loop over all bits which
was not ideal.

This commit revises the function to optimize the loop over the bitset by
dealing with a few special cases: no set bits, a single set bit, and all
set bits. The idea is that the bitset is very likely going to be sparse
or very dense, and these special cases can be optimized.

Other changes:
- Adds serial version of the offset getting function for times when you
  need to call it in a parallel section
- Refactors counting code
- Adds a test for count/offset
- Renames the slow DynamicBitset test to free up the name for the actual
  dynamic bitset test