libgalois: fix DynamicBitset bug
RestoreTrailingBitsInvariant was constructing a mask of all zeroes in
the case that the size of the bitset was a mutliple of 64. This resulted
in masking off the entire last element in the vector.

Fix this bug and add some more unit tests that give broader coverage of
DynamicBitset and act as a regression test for this bug in particular.