build: Use canonical python path
For

    $BUILD_DIR/python_env.sh python $SRC_DIR/x.py

The same module is available through two import paths. One path through
the source directory and another through the build directory.

This normally doesn't cause an issue because the contents of the modules
are the same and at least one module will be found, but recent versions
of pytest (7.1.0 at least) check that modules are unique and can fail
with the following error

    [....]/site-packages/_pytest/config/__init__.py:625: in _importconftest
        assert mod not in mods

when a module like conftest.py is available multiple times.

Fix this by using the build directory as the source (pun intended) of
tests.